refactor: #332 MovingStatusUseCase 프로토콜 분리
